[TESTS] Use CMake to build and run tests
Tests are now managed with cmake, with the intention of eventually
moving the whole build system to it. The root Makefile still has
a test target that has been adjusted to call into the appropriate
cmake commands.

Also, GTest library is now downloaded on demand the first time the
tests are run, so we can delete it from the repository.